Fix doubleclick-on-login.
The actual reason for doubleclick on login is that edX loses the 'next' URL parameter during the procedure,
and django_cas supplies the HTTP referer by default if it sees no next_page parameter.
During login-while-logged-out, the referer is the CAS server URL, so it behaves as if a
doubleclick is required to get through, which is mighty annoying the users.

For extra safety, session-timeout middleware also needs to handle the same situation.